+++ lib/response.ml
@@ -1,6 +1,3 @@
-type t = {
-  status : int;
-  body : Yojson.Safe.t option;
-}
+type t = { status : int; body : Yojson.Safe.t option }
 
 let make ?body ~status () = { status; body }